+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [6.1.0] - 2019-09-08
+
+- move poll wrapper on top of select (only used on Windows) to the ix namespace
+
 ## [6.0.1] - 2019-09-05
 
 - add cobra metrics publisher + server unittest